Fix audit-service schema mismatch + make notifications async
- WorkLogAuditRecord: BigInteger → Integer for plannedHours/bookedHours
  (Hibernate maps BigInteger to NUMERIC but the migration column is INTEGER;
  Integer matches int4 exactly and hours values fit comfortably in int range)
- TaskEventConsumer: convert BigInteger from event to int via .intValue()
- AuditConsumerIT: update assertions to compare plain int instead of BigInteger
- NotificationService.notify(): annotate with @Async so the Kafka consumer
  thread returns immediately and email + user-lookup run in the background

+++ audit-service/src/main/java/com/demo/audit/model/WorkLogAuditRecord.java
@@ -5,7 +5,6 @@ import com.demo.common.event.TaskChangeType;
 import jakarta.persistence.*;
 import lombok.*;
 
-import java.math.BigInteger;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -39,10 +38,10 @@ public class WorkLogAuditRecord {
     private WorkType workType;
 
     /** Null for DELETED events. */
-    private BigInteger plannedHours;
+    private Integer plannedHours;
 
     /** Null for DELETED events. */
-    private BigInteger bookedHours;
+    private Integer bookedHours;
 
     private Instant changedAt;   // when the change happened in task-service
     private Instant recordedAt;  // when audit-service stored this record
